Drop comrak backend
The comrak backend is a lot easier to work with, but the extra
dependencies are not worth it when the pulldown-cmark backend is more tested.

Having both comrak and pulldown-cmark used to be ok, but with the
addition of the macro "backend" the comrak backend has to go for
maintenance sake.